check_for_level_transfer: Convert float to int
Emscripten complains about the implicit conversion.

+++ src/game/check_for_level_transfer.cpp
@@ -20,7 +20,10 @@ void game::check_for_level_transfer()
 
         if(change)
         {
-            point = {p.area.x, p.area.y};
+            point = {
+                static_cast<int>(p.area.x),
+                static_cast<int>(p.area.y)
+            };
             break;
         }
     }
